Added Daily Codewars Challenge
Provided a function named find that accepts two parameters in the following order: array, element, return the index of the element if found and "Not Found" otherwise. Shorten the code as much as possible in order to meet the strict character count (no more than 161) You may assume that all array elements are unique.