Added Daily Codewars Challenge
Complete the function howManydays. It accepts 1 parameter month, which means the month of the year. Different months have a different number of days as shown in the table below. Return the number of days that are in a month. There is no need for input validation: month will always be greater than 0 and less than or equal to 12.