Added Daily Codewars Challenge
Coding in function grabDoll. the function accepts 1 parameter: dolls. it's a string array, a list of some dolls.

You need traverse dolls by using for loop. If the element is "Hello Kitty" or "Barbie doll", you should push it to a bag(bag is an array, I've defined in the function); if it's other strings, we should use continue to skip it.

When the bag has three elements, the bag is full. You should use break jump out the loop; If the bag is not full, you should traverse dolls until the last element.

Return the bag after for loop is finished.