Added Daily Codewars Challenge
There was a test in your class and you passed it. Congratulations!
But you're an ambitious person. Your goal is to see if you're better than the average student in your class.

You receive an array of your peers' test scores. Now calculate the average and compare your score!

Return True if you're better, else False!