Added Daily Codewars Challenge
Create a function that will trim a string (the first argument given) if it is longer than the requested maximum string length (the second argument given). The result should also end with "..."

If the string is smaller or equal than the maximum string length, then simply return the string with no trimming or dots required.


If the requested string length is smaller than or equal to 3 characters, then the length of the dots is not added to the string length.